Refactored Roles import/export step.
